+++ src/components/cartContentSection/CartContentSection.tsx
@@ -1,5 +1,5 @@
 import { CartItemCardList } from "../cartItemCardList/CartItemCardList";
-import { CartSummary } from "../cartSummary/CartSummary";
+import { CartSummary } from "../cartSummary/cartSummary/CartSummary";
 import {
   StyledCartContentSection,
   StyledContentContainer,
