+++ src/components/cartItemCard/CartItemCard.tsx
@@ -2,7 +2,7 @@ import { useSetRecoilState } from "recoil";
 import { deleteCartItem, patchCartItemQuantityChange } from "../../api";
 import { cartItemsState } from "../../recoil/atoms";
 import { CartItem } from "../../types";
-import { ActionButton } from "../button/ActionButton";
+import { ActionButton } from "../button/actionButton/ActionButton";
 import {
   StyledCartItemCard,
   StyledCartItemCardHeader,
