+++ src/pages/cartPage/CartPage.tsx
@@ -1,12 +1,12 @@
 import { useEffect } from "react";
 import { useRecoilValue, useSetRecoilState } from "recoil";
-import { getCartItems } from "../api";
-import { ConfirmButton } from "../components/button/ConfirmButton";
-import { CartContentSection } from "../components/cartContentSection/CartContentSection";
-import { CartHeader } from "../components/cartHeader/CartHeader";
-import Header from "../components/header/Header";
-import { cartItemsState } from "../recoil/atoms";
-import { categoryCountState } from "../recoil/selector";
+import { getCartItems } from "../../api";
+import { ConfirmButton } from "../../components/button/confirmButton/ConfirmButton";
+import { CartContentSection } from "../../components/cartContentSection/CartContentSection";
+import { CartHeader } from "../../components/cartHeader/CartHeader";
+import Header from "../../components/header/Header";
+import { cartItemsState } from "../../recoil/atoms";
+import { categoryCountState } from "../../recoil/selector";
 import { StyledCartPage } from "./CartPage.styled";
 
 export const CartPage: React.FC = () => {
