+++ src/pages/orderConfirmationPage/OrderConfirmationPage.tsx
@@ -1,18 +1,18 @@
+import { useEffect } from "react";
 import { useRecoilState, useRecoilValue } from "recoil";
-import { ConfirmButton } from "../components/button/ConfirmButton";
-import Header from "../components/header/Header";
+import { getCartItemCounts } from "../../api";
+import { ConfirmButton } from "../../components/button/confirmButton/ConfirmButton";
+import Header from "../../components/header/Header";
+import { cartItemsCountState } from "../../recoil/atoms";
+import { categoryCountState, totalPriceState } from "../../recoil/selector";
 import {
-  StyledConfirmatioinPagePriceContainer,
+  StyledConfirmationPagePriceContainer,
   StyledConfirmationPage,
   StyledConfirmationPageDescription,
   StyledConfirmationPagePrice,
   StyledConfirmationPageSubTitle,
   StyledConfirmationPageTitle,
 } from "./OrderConfirmationPage.styled";
-import { categoryCountState, totalPriceState } from "../recoil/selector";
-import { getCartItemCounts } from "../api";
-import { useEffect } from "react";
-import { cartItemsCountState } from "../recoil/atoms";
 
 export const OrderConfirmationPage: React.FC = () => {
   const totalPrice = useRecoilValue(totalPriceState);
@@ -44,14 +44,14 @@ export const OrderConfirmationPage: React.FC = () => {
           </span>
           <span> 최종 결제 금액을 확인해 주세요.</span>
         </StyledConfirmationPageDescription>
-        <StyledConfirmatioinPagePriceContainer>
+        <StyledConfirmationPagePriceContainer>
           <StyledConfirmationPageSubTitle>
             총 결제 금액
           </StyledConfirmationPageSubTitle>
           <StyledConfirmationPagePrice>
             {totalPrice.toLocaleString()}원
           </StyledConfirmationPagePrice>
-        </StyledConfirmatioinPagePriceContainer>
+        </StyledConfirmationPagePriceContainer>
       </StyledConfirmationPage>
       <ConfirmButton text="결제하기" backgroundColor="rgba(190, 190, 190, 1)" />
     </>
