+++ src/router.tsx
@@ -1,7 +1,7 @@
 import { createBrowserRouter } from "react-router-dom";
 import App from "./App";
-import { CartPage } from "./pages/CartPage";
-import { OrderConfirmationPage } from "./pages/OrderConfirmationPage";
+import { CartPage } from "./pages/cartPage/CartPage";
+import { OrderConfirmationPage } from "./pages/orderConfirmationPage/OrderConfirmationPage";
 
 const router = createBrowserRouter([
   {
